Rename the guess action creator
